+++ opr/primer.py
@@ -15,6 +15,7 @@ from .params import PRIMER_ATTRIBUTE_NOT_COMPUTABLE_ERROR
 from .functions import molecular_weight_calc, basic_melting_temperature_calc, salt_adjusted_melting_temperature_calc, gc_clamp_calc
 from .functions import e260_ssnn_calc
 
+
 class MeltingTemperature(Enum):
     """Mode used to calculate the Melting Temperature of the Primer accordingly."""
 
@@ -101,7 +102,9 @@ class Primer:
         :type other_primer: Primer
         :return: result as bool
         """
-        return self._sequence == other_primer._sequence
+        if isinstance(other_primer, Primer):
+            return self._sequence == other_primer._sequence
+        return False
 
     def __add__(self, other_primer):
         """
@@ -291,7 +294,7 @@ class Primer:
             self._double_runs = counts
             self._computed["double_runs"] = True
         return self._double_runs
-    
+
     @property
     def E260(self):
         """