+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.4] - 2025-03-18
 ### Added
 - Salt-adjusted melting temperature calculation
 - `__contains__` overload
@@ -56,7 +57,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `complement` method
 - `reverse` method
 
-[Unreleased]: https://github.com/openscilab/opr/compare/v0.3...dev
+[Unreleased]: https://github.com/openscilab/opr/compare/v0.4...dev
+[0.4]: https://github.com/openscilab/opr/compare/v0.3...v0.4
 [0.3]: https://github.com/openscilab/opr/compare/v0.2...v0.3
 [0.2]: https://github.com/openscilab/opr/compare/v0.1...v0.2
 [0.1]: https://github.com/openscilab/opr/compare/0baa8dd...v0.1